SceneManager: refactor
